move handler_arg fixture to conftest
