+++ jupyterlab_nvdashboard/tests/test_accelerator_handlers.py
@@ -19,21 +19,6 @@ from jupyterlab_nvdashboard.handlers import AcceleratorStatusHandler
 
 
 @pytest.fixture
-def handler_args():
-    """Fixture to provide mock application and request for handler init."""
-    with (
-        patch("tornado.web.Application") as mock_application,
-        patch("tornado.httputil.HTTPServerRequest") as mock_request,
-    ):
-        # Mock the settings to return appropriate values
-        mock_settings = {
-            "base_url": "/",
-        }
-        mock_application.settings = mock_settings
-        yield mock_application, mock_request
-
-
-@pytest.fixture
 def authenticated_handler(handler_args):
     """
     Create an AcceleratorStatusHandler with authentication properly mocked.
